Allow to specifcy kernel path before submitting

+++ submit-functional-tests.sh
@@ -4,12 +4,16 @@
 TOKEN=$(cat ./env/jenkins.token)
 HOST=$(cat ./env/jenkins.host)
 
+if [ -z "$KERNEL" ]; then
+	echo No '$KERNEL' variable set
+	exit 1
+fi
+
 # Run all functional tests
 for job in $(ls ./lava-jobs-functional/*.yaml);
 do
 	echo Submitting $job...
-	cp $job _submit.yaml
-	
+	sed "s/KERNEL/${KERNEL//\//\\/}/g" $job > _submit.yaml
 	lavacli \
 	--uri http://hyeyoo:$TOKEN@$HOST/RPC2/ jobs submit _submit.yaml
 	rm _submit.yaml
